Moved ExternalRecombiner to FastJet add-on
